// src/app/shared/model/asset.model.ts
export class Asset {
  id: string;
  city:string;
  company_name: string;
  country: string;
  custom_field: any;
  email_perso: string;
  email_pro: string;
  facebook_url: string;
  first_name: string;
  full_name: string;
  last_name: string;
  linkedin_url: string;
  location: string;
  middle_name: any;
  personal_phone_number: string;
  role: string;
  school: string;
  standard_phone_number: string;
  status_of_email: string;
  twitter_url: string;
  image_url: string;
  list_id: string;

  constructor() {}

}